SignPages: Moved signatures_for_keyid to gpgmh
