+++ keysign/gpgmh.py
@@ -278,6 +278,46 @@ def sign_keydata_and_encrypt(keydata):
 
 
 
+
+from monkeysign.gpg import Keyring
+def parse_sig_list(text):
+    '''Parses GnuPG's signature list (i.e. list-sigs)
+    
+    The format is described in the GnuPG man page'''
+    sigslist = []
+    for block in text.split("\n"):
+        if block.startswith("sig"):
+            record = block.split(":")
+            log.debug("sig record (%d) %s", len(record), record)
+            keyid, timestamp, uid = record[4], record[5], record[9]
+            sigslist.append((keyid, timestamp, uid))
+
+    return sigslist
+
+
+def signatures_for_keyid(keyid, keyring=None):
+    '''Returns the list of signatures for a given key id
+    
+    This will call out to GnuPG list-sigs, using Monkeysign,
+    and parse the resulting string into a list of signatures.
+    
+    A default Keyring will be used unless you pass an instance
+    as keyring argument.
+    '''
+    if keyring is None:
+        kr = Keyring()
+    else:
+        kr = keyring
+
+    # FIXME: this would be better if it was done in monkeysign
+    kr.context.call_command(['list-sigs', keyid])
+    siglist = parse_sig_list(kr.context.stdout)
+
+    return siglist
+
+
+
+
 ## Monkeypatching to get more debug output
 import monkeysign.gpg
 bc = monkeysign.gpg.Context.build_command
@@ -287,7 +327,3 @@ def build_command(*args, **kwargs):
     log.debug("Building cmd: %s", ' '.join(["'%s'" % c for c in ret]))
     return ret
 monkeysign.gpg.Context.build_command = build_command
-
-
-
-
